+++ examples/list-example.html
@@ -4,13 +4,10 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>HTML-to-Vega-Lite List Example</title>
-    <!-- Vega, Vega-Lite, and Vega-Embed dependencies -->
     <script src="https://cdn.jsdelivr.net/npm/vega@5"></script>
     <script src="https://cdn.jsdelivr.net/npm/vega-lite@5"></script>
     <script src="https://cdn.jsdelivr.net/npm/vega-embed@6"></script>
-    <!-- Built library bundle (exposes HTMLVegaLite global) -->
     <script src="../dist/index.umd.js"></script>
-    <!-- Page styles for layout and code blocks -->
     <style>
       body {
         font-family: Arial, sans-serif;
