+++ src/strategies/implementations/heading-tag-strategy.ts
@@ -19,56 +19,13 @@ export class HeadingTagStrategy extends BaseTagStrategy {
   }
 
   public parse(context: ParseContext): ParsedOutput {
-    const { isClosingTag, attributes, currentStyle, segments, remainingParts, currentIndex } = context;
+    const { isClosingTag } = context;
     
     if (isClosingTag) {
-      // For closing heading tags, add line break if there's more content
-      const newSegments: typeof segments = [];
-      
-      if (this.hasMoreContent(remainingParts, currentIndex)) {
-        newSegments.push({
-          text: '\n',
-          ...currentStyle
-        });
-      }
-
-      return {
-        newSegments,
-        updatedStyle: currentStyle, // Will be ignored since we're popping
-        pushStyleToStack: false,
-        popFromStyleStack: true,
-        errors: []
-      };
+      return this.handleBlockLevelClosingTag(context);
     }
 
-    // For opening tags, validate and add line break if needed
-    const errors: string[] = [];
-    if (this.validateAttributes) {
-      const validation = this.validateAttributes(attributes);
-      if (!validation.isValid) {
-        errors.push(...validation.errors);
-      }
-    }
-
-    const newSegments: typeof segments = [];
-    
-    // Add line break before heading if needed
-    if (this.needsLineBreak(segments)) {
-      newSegments.push({
-        text: '\n',
-        ...currentStyle
-      });
-    }
-
-    const newStyle = this.applyStyle(currentStyle, attributes, context.tagName);
-
-    return {
-      newSegments,
-      updatedStyle: newStyle,
-      pushStyleToStack: true,
-      popFromStyleStack: false,
-      errors
-    };
+    return this.handleBlockLevelOpeningTag(context);
   }
 
   public getTagNames(): string[] {
