+++ src/strategies/implementations/highlight-tag-strategy.ts
@@ -1,51 +1,18 @@
-import { TextStyle, ParseContext, ParsedOutput } from '../../types';
+import { TextStyle } from '../../types';
 import { BaseTagStrategy } from '../interfaces/base-tag-strategy';
 
 /**
  * Strategy for highlight tags: <mark>
+ * 
+ * This strategy applies highlight styling to text.
+ * Uses the default parse() implementation from BaseTagStrategy.
+ * 
  */
 export class HighlightTagStrategy extends BaseTagStrategy {
   public applyStyle(currentStyle: TextStyle, attributes: string, tagName?: string): TextStyle {
     return {
       ...currentStyle,
       color: '#212529' // Highlight text color
-      // We may further add below properties if needed:
-      // backgroundColor: '#ffff00',
-      // padding: '0.125rem'
-    };
-  }
-
-  public parse(context: ParseContext): ParsedOutput {
-    const { isClosingTag, attributes, currentStyle } = context;
-
-    if (isClosingTag) {
-      // For closing tags, just pop from style stack
-      return {
-        newSegments: [],
-        updatedStyle: currentStyle, // Will be ignored since we're popping
-        pushStyleToStack: false,
-        popFromStyleStack: true,
-        errors: []
-      };
-    }
-
-    // For opening tags, validate and apply styling
-    const errors: string[] = [];
-    if (this.validateAttributes) {
-      const validation = this.validateAttributes(attributes);
-      if (!validation.isValid) {
-        errors.push(...validation.errors);
-      }
-    }
-
-    const newStyle = this.applyStyle(currentStyle, attributes, context.tagName);
-
-    return {
-      newSegments: [],
-      updatedStyle: newStyle,
-      pushStyleToStack: true,
-      popFromStyleStack: false,
-      errors
     };
   }
